requirements.txt
dotenv
asgiref==3.2.7
beautifulsoup4==4.9.1
boto3==1.13.24
botocore==1.16.24
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.8
django-allauth==0.41.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-rest-auth==0.9.5
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docutils==0.15.2
html5lib==1.1
idna==2.9
jmespath==0.10.0
lxml==4.5.2
Markdown==3.2.1
oauthlib==3.1.0
Pillow==7.1.1
psycopg2-binary==2.8.5
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.14.0
sorl-thumbnail==12.6.3
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.8
webencodings==0.5.1
